+++ examples/control_section.py
@@ -1,7 +1,7 @@
 import xarray as xr
-from xrviz.control.section import Control
+from xrviz.control import Control
 
-ds = xr.open_dataset("../xrviz/sample_data/great_lakes.nc");
+ds = xr.open_dataset("../xrviz/sample_data/great_lakes.nc")
 
 control = Control(ds)
 control.show()
