+++ examples/display_section.py
@@ -1,5 +1,5 @@
 import xarray as xr
-from xrviz.display.section import Display
+from xrviz.display import Display
 
 ds = xr.open_dataset("../xrviz/sample_data/great_lakes.nc")
