+++ xrviz/control.py
@@ -1,8 +1,8 @@
 import panel as pn
 import xarray as xr
-from ..sigslot.base import SigSlot
-from ..display.section import Display
-from ..describe.section import Describe
+from .sigslot import SigSlot
+from .display import Display
+from .describe import Describe
 
 
 class Control(SigSlot):
@@ -11,13 +11,13 @@ class Control(SigSlot):
         self.data = data
         self.displayer = Display(self.data)
         self.describer = Describe(self.data)
-        
+
         self._register(self.displayer.select, "selection")
 
         self.connect("selection", self.describe)
-        
+
         self.panel = pn.Row(self.displayer.panel, self.describer.panel)
-        
+
     def describe(self, _):
         selected_property = self.displayer.selected_property
         sub_property = self.displayer.selected_subproperty
