+++ xrviz/describe.py
@@ -3,7 +3,7 @@ import panel as pn
 import os
 import sys
 import xarray as xr
-from ..sigslot.base import SigSlot
+from .sigslot import SigSlot
 
 
 class Describe(SigSlot):
@@ -14,7 +14,7 @@ class Describe(SigSlot):
     ----------
     data: `xarray` instance: `DataSet` or `DataArray`
         datset is used to initialize the DataSelector
-    
+
     Attributes
     ----------
     panel: Displays the generated template
@@ -23,46 +23,47 @@ class Describe(SigSlot):
     def __init__(self, data):
         super().__init__()
         self.data = data
-        self.panel = pn.pane.HTML(style={'font-size': '12pt'}, width=400, 
-                                  height=100) 
+        self.panel = pn.pane.HTML(style={'font-size': '12pt'}, width=400,
+                                  height=100)
         self.panel.object = "Description Area"
-        self._template_load_path = os.path.join(os.path.dirname(__file__), 
+        self._template_load_path = os.path.join(os.path.dirname(__file__),
                                                 "templates")
         self._template_env = Environment(loader=FileSystemLoader(self._template_load_path))
         self._variable_template = self._template_env.get_template('variable.html')
         self._coordinate_template = self._template_env.get_template('coordinate.html')
         self._dimension_template = self._template_env.get_template('dimension.html')
         self._attribute_template = self._template_env.get_template('attribute.html')
-    
+
     def variable_pane(self, var):
         if var is not None:
-            variable_attributes = [(k, v) for k, v in self.data[var].attrs.items()]
+            variable_attributes = [(k, v)
+                                   for k, v in self.data[var].attrs.items()]
             return self._variable_template.render(variable_attributes=variable_attributes,
                                                   var=var)
         else:
             return self._variable_template.render(var=None)
-    
+
     def attribute_pane(self, attr):
         if attr is not None:
             attribute_description = self.data.attrs[attr]
-            return self._attribute_template.render(attribute=attr, 
+            return self._attribute_template.render(attribute=attr,
                                                    attribute_description=attribute_description)
         else:
             return self._attribute_template.render()
-    
+
     def coordinate_pane(self, coord):
         if coord is not None:
             return self._coordinate_template.render(coordinate=coord)
         else:
             return self._coordinate_template.render()
-    
+
     def dimension_pane(self, dim):
         if dim is not None:
-            return self._dimension_template.render(dimension=dim, 
+            return self._dimension_template.render(dimension=dim,
                                                    count=self.data.dims[dim])
         else:
             return self._dimension_template.render()
-    
+
     def setup(self, selected_property, sub_property):
         if selected_property == 'Attributes':
             self.panel.object = self.attribute_pane(sub_property)
