+++ xrviz/sigslot.py
