+++ xrviz/templates/coordinate.html
